Reorganizes ARTIS module structure and fixes imports
Moves ARTIS-related functionality into a dedicated module structure and updates import paths accordingly.

Reorganizes imports in the ARTIS data module for better readability and fixes docstring reference to use correct class name.

Adds missing newline at end of file to follow Python conventions.
