+++ tardis/io/model/readers/base.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pandas as pd
 
-from tardis.io.model.readers.artis import read_artis_density
+from tardis.io.model.artis.readers import read_artis_density
 from tardis.io.model.readers.cmfgen_deprecated import (
     read_cmfgen_composition,
     read_cmfgen_density,
